Retrieve list of Backup Items from Database
We are now able to retrieve the backup items directly from the database.
This means that we are going to be able to use the model for requesting
all items for either display or the actual backup process. It will be
expanded to be a list of tuples that will contain all the backup list
data: source, backup frequency, relevent settings, name of the source,
source type, anythign else needed. It also sorts the list as of now so
that it will be easier for us to read.