Remove Added, Returning list of items works, bugs fixed
The remove function allows us to remove a backup item
from our backup list. The return list now returns a dictionary
that contains a tuple of all data associated with each backup
item. Various bug fixes are included to help check if we
are making valied queries.

The remove function works through takign a name and a type
for a given node and finding it in the backup list. We then
just remove it from the tree. This works. We have to manually
save as I don't want users to feel that they must redo everything
incase they mess up. Working so far.

The returnlist function is goign to be an expandable dictionary
of tuples. We will use the text as the key because it is what the user
will see and I __THINK__ it will be easier to access via ttk.
Also, since the actual backup script won't care about names we can
just for loop over it and get all relevent information out of the tuple.
This is perfect for my purposes.

Bug fixes mainly focus around the checking to make sure we didn't get a
None element. Also there are a few try catch blocks added to allow us
to abort the project on the event that we have a corrupt xml scheme.
I want to try to catch those as soon as possible so that they aren't written.
If we can abort and not write our xml without it being clean then we
will save loss of data for the users who can't fix xml manually.

Signed-off-by: Barrett Lewis <musikal.fusion@gmail.com>