Using doxygen - fixed bumodel.py
Pushing the doxy information as well as the new bumodel with doxygen
style comments.

I have changed the name from Python Back Up to Python Bakup Utility:PyBakUP.

The new html documentation has been pushed with the project so that it is easy to access to those working on the code.
I know this might be premature but I want to use doxygen throughout the project since it is new ^_^