Updated the xml structure to make way more sense
The xml structure has been updated to be less error prone. Namely,
source and file have been moved to attributes of the bi tag.

The main problem was this, files and folders had unique tags. This made
no sense because we had to guess which was being backed up. The new
structure makes much more sense because tags are more meaningful.  I
tried to encapsulate meaningful things into their own tags so they can
be accessed as needed. I am not sure I am at an optimal set up yet but
it is orders of magnitude better than the last setup. Everything is
beginning to come together for the database interactions.

Signed-off-by: Barrett Lewis <musikal.fusion@gmail.com>