Encapsulated model, viewer has acces to instance
I have set up a more encapsulated version of the model class using the
module to simulate a singleton. The viewer is now accesing it.

I have set up the Model to simulate a private class (we are all adults
don't instantiate __Model). Instead the module has an accessor that uses
a global variable for the instance. Maybe a static member might work
better, either way not sure yet. To get the instance you just use
bumodel.GetInstance().

The viewer is now able to get the instance, I am contemplating setting
up a class for this, not sure yet.