First part of gui complete
The list view of items is now building properly.
I added an interface ability to buumodle, if it is the main file the ui
is utilized. Ran doxygen but it isn't completely up to date.

At this point we are able to use the db to build the list used by the
gui to display the backup items. The temporary ui for the model is now
only activated if the program is started from bumodel.py. Everything
seems to be working so far.

Signed-off-by: Barrett Lewis <musikal.fusion@gmail.com>