Most production actions done, lexer now uses toks
I added many of the actions for each of the productions. I had to split
some of the stuff up so that it would make more sense. I think we may
get away without any extra wrapper classes( the language is VERY simple)

The lexer was changed to return tokens instead of just returnign what
was typed.